Initial Commit
This was probably between 8 - 12 hours worth of work. There's quite a
bit of work that's still possible on the API, but all basic
functionality is present. The primray deficiencies right now are in
security and deployment. There's no concept of environments and cookies
are served plaintext with a simple user id, so its relatively simple to
impersonate other users by guessing ids.

In terms of features, the API needs some logging facilities. I've set up
sessions to have an expiration time but nothing touches that yet.

There are a lot of things I get essentially for free though! Automated
generation of client side JS functions to interact with the API.
Automated generation of API documentation with a little bit of extra
effort to specify samples and usages. Automated generation of a mock API
that spits out random data based on QuickCheck with a little bit of
extra effort (copy-paste from a blog post). All the guarantees that
strong static typing brings, primarily robustness.

A note about the persistence layer. I made a choice to use AcidState as
the persistence layer instead of a traditional database like Postgresql.
This was a trade-off. AcidState stores plain Haskell datatypes, so
there's no impedance mismatch between the database and the system.
However AcidState operates entirely in RAM, and doesn't provide all of
the things a database would provide like indexes and constraints. I
needed to do those myself. In general I think the lack of an impedance
mismatch wins out for smaller scale projects like this.

I used to have this broken up into QuestionsAPI and SessionsAPI, each
under their own endpoint. However it became clear that Questions did
much better being nested under Sessions, so QuestionsAPI was deleted and
SessionsAPI ate those endpoints. I kept around the individual API file
in case some future changes require truly separate endpoints, but it
could be merged into API.

Servant is the API framework I used here. It's a sophisticated framework
that pushes as much as possible into the type system. Fortunately I've
found that a lot of it doesn't require a deep knowledge of how it works
under the hood. At least I don't have that knowledge after working with
this project. By doing a lot of things in the type system Servant is
able to provide a lot of nice things for free, like automatic parsing
and handling of request data, client side javascript generation, mock
apis, and documenation generation. And it's one of the most concise
frameworks I've used. All in all I'm very impressed.