Generate Swagger spec.
This is basically the bare minimum. The spec could be improved by adding
more prose descriptions and making sure datatypes are encoded
accurately. I've noticed that Haskell data types that don't use record
syntax get encoded as arrays, which isn't great. There's also a
parameter in JoinReq that should be optional that seems to be listed as
required. All in all, a good starting point.