Project is built against a specific set of packages
This sets up the project so that it is always built against the same set
of packages. This has the effect of making it fully reproducible, so two
devs using Nix on two machines but with different versions of the
Nixpkgs repository will still get the same build. It also won't
interfere with existing software installed using Nix.