Use servant-aeson-specs to test JSON parsing and generation
The servant-aeson-specs library is pretty new, but introduces a lot of
convenience. I'm using it to automatically generate two sets of
QuickCheck tests for every datatype used in the API definition.

The first set of tests (apiRoundTripSpecs) simply tests that `toJSON
(fromJSON x) = x`. This ensures that no information is lost when
generating and parsing JSON.

The second set of tests (apiGoldenSpecs) tests that JSON generation does
not unintentionally change. It saves sample data and the JSON generated
for it, and checks that the current generation methods still generate
the saved data. This ensures that bugs aren't accidentally introduced
that change the JSON format.

The library is still pretty new, so a few caveats apply:

Currently, if built using Nix, the servant-aeson-specs is failing it's
tests, so I introduce an override to skip that check.

Additionally the library is missing some typeclass instances that
prevent it from handling my API definition (Capture, Headers, and
AuthProtect). I've introduced them here.

Finally, the library requires Eq, ToJSON, and FromJSON instances for
each datatype it's testing. I've introduced them here, even though some
of the FromJSON instances aren't currently being used. Now they'll be
written and consistent when they're needed.