Refactor SessionsAPI Handlers so (almost) all logic resides in the *Handler functions
This caused the *Handler functions to change return type to use ExceptT,
in order to allow for proper error throws. Without ExcepT and just using
IO, a subtle bug arises when doing something like:

```
answerQuestionHandler app (AnswererCookie  answererId) sessionId questionId sessionChannelDB = do
    q <- liftIO $ update app (AnswerQuestion questionId sessionId answererId)
    sendSessionMessage sessionChannelDB sessionId QuestionAnswered q
    return q
```

If liftIO were to fail, sendSessionMessage would still be executed. This
is not what we want.

The withError utility method also got a little simpler in that now it's
a simple translation of error type, instead of also a translation into
the ExceptT monad.

The changes to the type of fullSessionHandler required
corresponding changes in the WebsocketAPI module.

Finally, the logic for adding session cookies does not live in the
*Handler functions. That doesn't really seem to pertain to business
logic, but I could be persuaded. Mostly, it would make testing difficult
by merging the authentication and business layers.