Add User types and endpoint
Design notes:

There's a User and a FullUser type. This is to prevent accidental
exposure of user passwords. A developer must be intentional about using
the type that contains a user password. Looking through the code you'll
notice that all the API endpoints expose a User without a password.

Passwords are wrapped up in a newtype to restrict the operations that
can be performed on them. Important to note is that passwords can't be
compared for equality directly. This is because in the future a constant
time equality comparison method is likely to be included.

FullUsers are compared using reference equality. This stems from the
fact desire to not allow Passwords to be compared for equality. I could
maybe implement it to compare all the User fields, but I think reference
equality is probably good enough here.

Authentication is initially done via a login request and then an
encrypted cookie afterwards. There is no CSRF protection at the moment.

Code comments:

There are some pretty glaring security flaws that are spelled out more
fully in the code.
I needed to include a ToJSON instance for the LoginReq request body. I
don't want to do that as it poses a security risk by allowing those to
be serialized to JSON and sent on the wire by accident.
There aren't any password rules yet.