Remove old Answerer/Questioner authentication/authorization system.
This looks complicated, but the actual changes were pretty
straightforward. Change the data structures to the desired end shape
then follow the compiler errors until they're all fixed.

Notes:
I'm not satisfied with how the Migrations are organized.
This removes the ability to anonymously authenticate and interact with a
session.